requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
redis>=3.5.0
pydantic>=1.8.0
python-dateutil>=2.8.0
PyYAML>=5.4.0
click>=8.0.0

# API dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose>=3.3.0
passlib>=1.7.4

# Monitoring
prometheus-client>=0.11.0

# Logging
loguru>=0.5.0
